requirements.txt
python-telegram-bot==20.7
Flask==2.3.2
Pillow==10.2.0